Modelling/RottenFresh/requirements.txt
numpy
pandas
matplotlib
opencv-python
Pillow
python-dotenv
scikit-learn
shap
torch
torchvision
kagglehub
